Add create-server acceptance test and support code
This check-in is not complete; it will break acceptance tests.  The
problem is that I cannot run the test in full-quiet mode yet; I need to
support listing of images and flavors before I can do that.  That will
allow the acceptance test to choose a server flavor and OS image
appropriate for the acceptance test.