Finish acceptance test for CreateServer() function
Two problems remain -- server status is coming up blank, and
measurements of disk and RAM are coming up zeros.  Not sure why this is.
It used to work in Gorax.  More comparative testing between Gorax and
Gophercloud is needed.

However, what is in this commit is sufficient to get the acceptance
tests running again, at least on Rackspace's infrastructure.  If you
attempt to run these tests on HP or other infrastructure, they will
fail for you.