Make portID not mandatory in floatingips
With the current code, it is not possible to allocate a floatingIP without
assigning it to a port. Some tools will require to allocate an IP first to
assign it later.

This is allowed in OpenStack API: http://developer.openstack.org/api-ref-networking-v2.html#layer3

This patch makes portID optional in gopher cloud