Replace struct based mapping by manual mapping
The old way does not allow to handle updates correctly. When a nullable
field is set and we want to remove the value we need to be able to set
a null value in the json request body. For instance this happen in
firewall rules for field source_ip_address (among others).