Make client return error on JSON decoding error
If the response is not valid JSON, the request simply returns an
empty body. If the user is expecting the result to be JSON and its
not, we should signal that an error has occured.

This patch also includes fixes to tests that fail due to the
error that bubbles up.