Fix NewClient to not remove relative paths from IdentityBase
gophercloud.provider_client.go struct ProviderClient describes IdentityBase as
follows: "IdentityBase is the base URL used for a particular provider's
identity service - it will be used when issuing authenticatation requests. It
should point to the root resource of the identity service, not a specific
identity version."

Currently gophercloud.openstack.client.go func NewClient strips endpoints with
a non-empty relative path. For example, the IdentityBase returned for endpoint:

  http://example.com/foo

is:

  http://example.com/

when it should be:

  http://example.com/foo/

This change corrects NewClient to correctly handle such cases.

Unit tests are provided in openstack.test_client.go

[1] gophercloud.openstack.client.go

Fixes-Issue: #577