Generate shell scripts in the Magnum credentials bundle (#64)
* Include the bay's docker client version

* Generate swarm compatible shell scripts in the credentials bundle

* Don't use a fixed/named set of scripts in the credentials bundle.
  Allow each coe type to use a different set of scripts

* Generate kubernetes compatible shell scripts in the credentials bundle

* Use carriage returns in shell scripts on Windows

Without this, when a Windows user opens one of the scripts in a standard
editor, like Notepad.exe, all the lines will be displayed on a single line.

* Consolidate bay and certificate bundle acceptance tests

It takes a long time to create a bay, and the bundle test originally
relied on one already existing. This moves the bundle test into the bay
test so that it's quicker and less fragile.