Server Metadata Update
This commit updates the Server Metadata so it can be parsed and
translated using JSON tags.

The Server Metadata result has been changed to map[string]string to
match the request.

A unit test was added to verify the above and a piece of metadata has
been added to the server creation acceptance test.