Update some docs for some nits found (#286)
* A few small doc clarifications

First of all, remove reference to HP as an OpenStack provider and
replace with some other examples of folks who are actually running
OpenStack Public Clouds.

Then add a few notes clarifying tenant/project usage and the fact that
domain applies to both.

* Change my-openstack.com to openstack.example.com

RFC 2606 defines example.com as a reserved domain to be used in tests
and examples. It is guaranteed to not be actually resolvable. Update
the uses of my-openstack.com to use example.com instead.

  https://tools.ietf.org/html/rfc2606

* Change "cluster" to "cloud"

The general term for an installation of OpenStack is cloud, at least
when talking to end users.

* Address review comments