Added Type into ServiceClient to Distinguish Different Services
All ServiceClient methods insert the OpenStack Nova HTTP Microversion header ("X-OpenStack-Nova-API-Version") into HTTP requests even though the request may be to a different service than Nova. This is incorrect.

That's why a mapping of serviceType to HTTPMicroversionHeader is added. Default value (empty string) is mapped to OpenStack Nova HTTP Microversion header because of backward compatibility.

The serviceType constants will be made public as soon as it will be needed.