Endpoint Locator in addition Returns EndpointExtraInfo struct
Data as service type, microversion, service name, etc. are needed in ServiceClient struct because clients of services like Nova Compute or Manila need to insert an extra HTTP header which name depends on the service name and its value contains the service microversion.
Data as service type, microversion, service name, etc. are not returned from the V2EndpointURL and V3EndpointURL functions even though they are available in these functions

That's why the EndpointExtraInfo struct was added into the EndpointLocator type return values. The EndpointExtraInfo struct is not used yet, is kept empty and is being ignored.