Jrperritt's Review Comments
Cleans up wine-induced typos.  Updated the unmarshall code, seems the integer case is
unnecessarily as go when unmarshalling a number defaults to float64, so has been
deleted.  As we default to 0.0 for total and free capacities I've folded the unknown
case into the default.  Testing code has been updated to include an 'integer' in the
JSON to ensure the assumptions hold.