Add Cinder V3
This just creates base V3 support for Cinder.  For all intents V3.0
of Cinder is in fact V2, so we don't need to do much with the existing
volumes or snapshots right now.  We do want to add microversion support
though so we can get shiny new things like the Attachments API.