Add Cinder V3 and MicroVersion Support
This change pulls the existing V2 implementations into V3.  Note that
the base V3 version (3.0) is a direct map back to V2.  As a result to
use the base V3 endpoint for Cinder we can don't need to change anything
other than the URL.  Also clean up some simple lint errors in
requests.go

One thing that the base V3 version does buy us here though is
Microversions (for better or worse).  So, this change also adds in the
ability to specify MicroVersion in the additional-header of the
ServiceClient.  This will be important going forward and so we can now
add things like the new standalone Attachment API.

Volume V3 just points to V2 unless a microversioned call is issued:
    https://github.com/openstack/cinder/blob/master/cinder/api/v3/volumes.py#L25
    https://github.com/openstack/cinder/blob/master/cinder/api/v3/volumes.py#L50

An example of a micro versioned call (not implemented in gophercloud):
    https://github.com/openstack/cinder/blob/master/cinder/api/v3/volumes.py#L212

Even the base V3 view builder just utilizes V2 (only adding additional
info for specific versions):
    https://github.com/openstack/cinder/blob/master/cinder/api/v3/views/volumes.py#L19
    https://github.com/openstack/cinder/blob/master/cinder/api/v3/views/snapshots.py#L19

Note that this change is part of a series that will add Attachment
support.