Methods for Parsing and Comparing Microversions
Some Manila service API requests are supported from a particular microversion. Therefore, it will be needed to compare two different microversions in order to find out if an API request is supported or not.

That's why the a data type Microversion with methods for parsing and comparing microversions is added.