Networking v2: Fix Port and Security Groups
This commit fixes the way security groups are handled during port
creation and updating.

If no SecurityGroups are specified during Create, then
`security_groups` is not passed in the request body. The default
behavior for this is to have the default security group applied to
the port.

If a pointer to an empty string slice is passed, then the port is
created with no security groups applied.

If no SecurityGroups are specified during Update, then no
modification is done to the port's security groups. If a pointer
to an empty string slice is passed, then all security groups are
removed.