Networking v2: Fix Port and Allowed Address Pairs
This commit fixes the way Allowed Address Pairs are handled
during port updating.

If no AllowedAddressPairs are specified during update, then
`allowed_address_pairs` is not passed in the request body. Existing
address pairs on the port will remain untouched.

If a pointer to an empty string slice is passed, then the address
pairs are removed from the port.