Remove autocomputed etag value
The autocomputed etag value should be optionnally set by the end user, not the lib. 
The computed value based on the content of the body is wrong in some usecases

e.g.: for multipart manifests, if set, the value must contain the MD5 checksum of the concatenated ETag values of the object segments.