Networking v2: Fix Extra DHCP Options Updates
This commit fixes updating of extra DHCP options on a Networking
port. The fix entails creating two new structs: one for create
and one for update. The update struct is able to take a string
pointer as a dhcp value so a nil/null string can be provided.
This nil/null value will cause a dhcp option to be remove if
requested.