Improve test fixture parmaeterization.
This allows us to more easily see problems with test cases by specifying
the expected inputs and outputs in the same place in code rather than
the implicit assumption that the fixture has the output expected for the
test case.

Another advantage is that this brings us that much closer to the ideal
of test fixtures consisting purely of a list of data structures which
can then be iterated over in a single test case function that glues all
the pieces together to validate that the tested code, given some input
data, produces the expected output data.