simple sed s/snapshot/backup on existing code
the backup and snapshot functionality of Cinder do much the same thing
with respect to how it interacts with the underlying objects ( above the
driver ). I've simply replaced all instances of [Ss]napshot with backup.
Future work will be tweaking the code to make it work properly with the
backup API of Cinder.