Compute v2: AllTenants Usage (#1125)
* Compute v2: AllTenants Usage

* Compute v2: Changing usage paging to LinkedPage

A microversion addition enabled paging support to usage results.
This does not add support for the microversion-added parameters,
but changes the pagination implementation to support them in the
future.

* Compute v2: Fixing usage paging

The usage API does not work well when using the AllPages method for two
reasons:

The format of the result is not a standard "list" format. A portion of
the result is a list, but the result also contains non-list data that is
not captured with AllPages.

In addition, some of this data is a calculation done on the usage within
the single page. The calculation then changes from page to page and this
is also not captured with AllPages.