core: Revert setting reauth func to nil
This commit reverts PR #1192. The goal of #1192 was to set the reauth
function to nil when performing reauthentication in order to prevent an
infinite loop. However, this was not done in a concurrent-safe way.
Having reviewed the use-case in more detail, I feel that scenarios in
which this will happen are more site-specific rather than an issue with
Gophercloud and a custom reauth function should be used.

Included in this PR is a simple unit test which verifies that a custom
reauthentication function can be used for these scenarios. However, it's
important to not use this example in production. Rather, users should
use the example provided in the FAQ.