add gophercloud.AuthResult (#1226)
* add gophercloud.AuthResult

The API design in this commit is (again) wildly different from the
various proposals that I came up with in #1141. Let's see what Joe
thinks of this.

* replace IsAnAuthResult by the more plausible ExtractTokenID

Also, since the first argument of SetTokenAndAuthResult is now
redundant, remove it and rename the method to SetTokenFromAuthResult
accordingly.

* address review comments

Mostly style. SetTokenFromAuthResult is renamed back to
SetTokenAndAuthResult to make it clear that client.authResult is set by
it.