Support Magnum resize API
Recently in Magnum we introduced a new API v1/clusters//actions/resize [1]
to enable the capability that consumer can resize the node count of the
cluster and meanwhile the API allows to pass in the node instance ID
so as to delete particular nodes. That's very useful for the auto-healing
and auto-scaling cases which Magnum will leverage k8s cluster Autoscaler
to achieve that.

Without this new API, the consumer has to call Heat API firstly to remove
the particular to remove the nodes and then call Magnum API again to
update the node count, which is not ideal obviously. Hence the new API
was proposed. More information please refer [2] [3].

[1] https://review.openstack.org/638572
[2] https://review.openstack.org/631378
[3] kubernetes/autoscaler#1690