Object Storage v1: Enable removal of container metadata
This commit enables metadata to be removed from a container. Gophercloud
internally strips empty strings from headers to enable deletion of
headers before they are sent in the request. This conflicts with the
ability to actually send an empty string when needed. Thereforer, a new
UpdateOpt has been added that sends the special X-Remove header with a
dummy value of "remove".