Baremetal introspection: correct handling of LLDP data (#1548)
* Baremetal introspection: correct handling of LLDP data

The LLDP data in the inventory doesn't appear as a JSON object, but as
a JSON array of [type, value] pairs such as:

[[1, "04112233aabbcc"], [2, "07373334"], ...]

(There can be multiple fields with the same type, which is why they are
not returned as a JSON object.)

There is a place where the raw LLDP data is processed into a JSON object
- under the `lldp_processed` key in the top-level `all_interfaces` list
(i.e. not in the inventory), here represented by BaseInterfaceType.

This patch corrects handling for LLDP data in the inventory (i.e. in
InterfaceType) and adds handling for it in BaseInterfaceType

Signed-off-by: Zane Bitter <zbitter@redhat.com>

* Baremetal introspection: Change Lldp to LLDP

Coding style prefers all caps for initialisms.

Signed-off-by: Zane Bitter <zbitter@redhat.com>

* Baremetal introspection: Check LLDP data in tests

Signed-off-by: Zane Bitter <zbitter@redhat.com>

* Baremetal introspection: Add tests for LLDP parse errors

Add tests for handling of invalid data in the code that deserialises
LLDP TLV data from JSON.

Signed-off-by: Zane Bitter <zbitter@redhat.com>