Block Storage: Backups (#1552)
* inital creation of v2/backups

cp -a snapshots backups

* simple sed s/snapshot/backup on existing code

the backup and snapshot functionality of Cinder do much the same thing
with respect to how it interacts with the underlying objects ( above the
driver ). I've simply replaced all instances of [Ss]napshot with backup.
Future work will be tweaking the code to make it work properly with the
backup API of Cinder.

* mend

* added a couple of options for Backups

backups.List() works, but .Create() does not. I keep getting HTTP 400.
Then, I get the same for snapshots.Create() which should work ( because
I haven't messed with it ).

Trying blockstorge/v3

* inital copy and replace of blockstorage/v3/backups

same as v2, "cp -a" and "sed -i -e "s/snapshot/backup/g"

* v3 backup works to list and create.

very much still a work in progress.

* Block Storage: Move backups to extensions

* Block Storage: clean up backups

* Block Storage: Add backup acceptance tests

* Block Storage: Microversion results for backups