Load Balancer v2: Remove Health Monitor Validation (#1640)
Octavia now supports defaults for when the url_path and expected_codes
are not supplied in the request body. This commit removes the local
validation for these fields to allow the API defaults to be used.

Users working with earlier versions of Octavia will need to be mindful
that they cannot rely on defaults being provided by the API and
therefore must provide explicit values.