Compute v2: Move ServerTagsExt to servers.TagsExt (#1760)
This commit moves the tags extension's ServerTagsExt to
the servers package. While this is slighly nuanced, the tags
extension can be thought of to work directly on a server's
tags and the ability to extract a server's tags can be part
of the servers package since the original ExtractTags
microversion support already exists.