Containerinfra: acceptance test timeout changes and fixes (#1794)
* Add WaitForTimeout to acceptance tools

To be used in situations where a 300 second timeout is
not long enough or otherwise unsuitable.

The WaitFor function is changed to simply call WaitForTimeout
and pass the same 300 second timeout, keeping the same behaviour.

* Set timeout for cluster creation in containerinfra acceptance tests

Keeps the same timeout as previously by default, but can be changed
for cluster templates that require a longer timeout.

* Check cluster creation error in containerinfra CRUD acceptance

* Use integer for replacing node count in containerinfra acceptance tests

Required for compatibility with Magnum APIs running on Python 3.
See https://github.com/gophercloud/gophercloud/issues/1458
and https://github.com/gophercloud/gophercloud/pull/1463

* Use overlay2 for template storage driver in contianerinfra acceptance

With the storage driver set to devicemapper the magnum API did not
accept clusters created with the template, as docker_volume_size
was not set. Instead of setting the volume size, overlay2 can be used
with no volume and is the preferred storage driver.