BlockStorage: apiversions (#458)
* BlockStorage: apiversions

* BlockStorage v1: Add deprecation note to apiversions

* BlockStorage: Fix URL when retrieving API version

* BlockStorage: Add acceptance test for apiversions

* BlockStorage: Base all Results on List

This commit removes the Get request and adds the ability to
extract a single API version via a ExtractAPIVersion function.

This is done to handle cases where older Block Storage v2 API
endpoints return a different style format than found in newer
releases of Cinder. Since the result of a List request to the
base API URL is the same across all releases, a consistent
response can be retrieved.