Add ip, ip6, and user_id to servers.ListOpts
As listed in https://docs.openstack.org/api-guide/compute/server_concepts.html#server-query:
* ip
* ip6 (New in version 2.5)
* user_id (New in version 2.83)

And the CLI documentation:

    --ip <ip-address-regex>
                          Regular expression to match IP addresses
    --ip6 <ip-address-regex>
                          Regular expression to match IPv6 addresses. Note that
                          this option only applies for non-admin users when
                          using ``--os-compute-api-version`` 2.5 or greater.
    --user <user>         Search by user (admin only) (name or ID)

Note the difference between the user_id docs. My understanding is that
it’s the same as ip6: the option was initially admin-only, and became
generally available with API 2.83 (this is what I have documented).

For #2035