Allow updating routers, without specifying the static routes
This is helpful when it is necessary to rename a router,
but it is not desired by the user to touch the static routes.

The OpenStack API just allows to omit the static routes (e.g.
verifyable with `openstack router set --name new_name --debug`).

This change makes sure that static routes parameter can be omitted with
gophercloud as well.

Also added a unit test.