Add missing bios interface fields to Ironic API
The bios interface fields are in the Ironic API but were missing from
the Ironic API documentation and not included in gophercloud. This
change adds in the bios interface to the node create, node list,
and node driver validation APIs.

For reference, these are the Ironic updates to the api-ref
documentation:
https://review.opendev.org/c/openstack/ironic/+/757353
https://review.opendev.org/c/openstack/ironic/+/791493
https://review.opendev.org/c/openstack/ironic/+/792210