Fix ErrUnexpectedResponseCode expected OK codes
When receiving a response from OpenStack HTTP response code is checked against a list of OK codes.
This list is either specific to the call being made or generic according to the HTTP verb.
In the latter scenario, when a wrong code is received, an error is returned that does not contains the actual list of expected OK codes.
It is misleading because you may think the list of accepted codes is not correctly defined.

Actual: 'Expected HTTP response code [] when accessing [DELETE http://127.0.0.1:35195], but got 200 instead OK'
Expected: 'Expected HTTP response code [202 204] when accessing [DELETE http://127.0.0.1:35195], but got 200 instead OK'