Pre-fetch the Octavia Amphora image
This will avoid issues that we currently have with DIB, until we figure
it out correctly. To figure it out correctly, we need to rework the
roles that we run for Devstack job to properly collect the logs.

This will be done in the future but not in this first effort.

For now, we'll download the 500MB+ image from RDO servers, which should
be faster and more reliable than building the image locally.