tests: Fix Go v1.14 "go vet"
We use "go get" to install test dependencies in Go v1.14. As a side
effect, `go.sum` was polluted and `golang.org/x/crypto` was bumped to a
version that is incompatible with Go v1.14.

With this change, installing test dependencies no longer changes the
state of the checked-out repository.