Add Github Workflow for checking Go API compatibility on PRs
It'll automatically set the semver label for a given PR.

Limitation: files under `/acceptance` or named `doc.go`, `fixtures.go`
and `fixture.go` which don't count as semver, will count in this new
job.

This is because we would need to add a commit that exclude these files,
then rebase the PR on top of that and run `go-apidiff`. This is too
complex for now and might be addressed later.