Add Github Workflow for checking Go API compatibility on PRs
It'll check API compatibility for each new PR.
If the check is compatible, the label `semver:minor` will be added.
If not, `semver:major` will be used.

Now there is no way to detect whether there is an actual API change or
not (ie make a diff between `minor` and `patch`).