objectstorage: Do not parse NoContent responses
Some Swift instances respond with HTTP status code 204 when asked to
list containers and there are no containers, or when asked to list
objects on an empty container. Before this patch, Gophercloud would
error because the header `content-type` is absent on the response.

With this patch, objectstorage responses with HTTP status code 204 are
immediately recognised as empty and their body is not read.