Make TestMTUNetworkCRUDL deterministic
It seems the MTU update API does not returned an object with the
`updated_at` field to it's new value. This can cause our test to fail if
we check the return of the Update function against the same object
retrieved from a Get call, as the `updated_at` field can differ.

Instead, just checked for the fields we change in this test.

Fix #2592