RELEASE.md: List PRs from last release
Stop relying on milestones. A release process based on milestones is
prone to error (because not all PRs were necessarily added to the right
milestone), and moreover it is not applicable to the cherry-pick
workflow.

Instead, this new release process uses `gh` to retrieve all PRs since a
given release.