identity: use AuthOptionsBuilder for v2 auth
Rather than creating a tokens2.AuthOptions inside the authentication
flow for identity v2, allow any structure that implements the
tokens2.AuthBuilderOptions interface to be used. The interface is
modified to add a CanReauth function like the v3 version has but
provides a default implementation. The gophercloud.AuthOptions already
implements this interface so there is no change to its API. This will
allow an out of tree identity v2 auth mechanism to be implemented,
just like the v3. fixes #2651. ref #1330.