[v1] identity: use AuthOptionsBuilder for v2 auth
Rather than creating a tokens2.AuthOptions inside the authentication
flow for identity v2, allow any structure that implements the
tokens2.AuthBuilderOptions interface to be used. The
gophercloud.AuthOptions already implements this interface so there
is no change to the API. To avoid any API change, add
AuthenticateV2Ext() as a version of AuthenticateV2() accepting the
interface. This will allow an out of tree identity v2 auth mechanism
to be implemented, just like the v3. fixes #2651. ref #1330.