tests: run MultiAttach with a capable Cinder Type
Starting from the Queens release, it is possible to attach a volume to
multiple hosts/servers, using a volume type with multiattach capability.

Cinder previously had a `multiattach` parameter that could be passed
during the volume creation. This parameter was recently removed [1], and
caused our jobs to fail.

This commit change the tests to no longer use the removed parameter.

Fixes #2657.

Co-Authored-By: Martin André <m.andre@redhat.com>

[1] https://github.com/openstack/cinder/commit/d4535c77493a7b362091b962f42f2613dea65dbe