[CI] Drop branch filters for 'Ensure Labels' and 'CodeQL' jobs
These jobs should run on all branches, so drop the branch filter that
were in place.

For the CodeQL job, it seems like the github action syntax does not allow for a simplified `on` syntax, while at the same time using a `on.schedule`, so drop the periodic that we weren't paying attention to anyway. We have enough signal with just the PR jobs.