workflow/semver-auto: rebase PR
When testing if a PR is API backward compatible or not, we need to test
against the base branch, not the head against the PR branched is based
on.

Otherwise it can lead to the case where a breaking change was merged
after that a PR was proposed, go-apidiff will fail because it'll test
the repo against base_ref.

The PR needs to be rebased on base_ref so the actual compatibility can be
verified.

If the PR can't be rebase for any reason, we'll return a failure and add
the label `semver:unknown`.

(cherry picked from commit 3e0071fcd67b3fb055c7634e1191a704dae8a56e)