ci/semver-auto: return a failure only when needed
When go-apidiff returns semver=major, it returns RC=1 which causes the
workflow to return an error while we don't want that, we just want the
label to be applied.

Instead, let's ignore that error so the label can be set later and add a
step to check if an error occured and semver is not major, then it's an
actual problem that we should check in the logs (e.g. an issue when
running go-apidiff).