acceptance/volumeattachments: use nova API for most of it
It is not recommended [1] to use volume attachment API from Cinder
directly when we want to test real user workflows, like creating a
server, a volume and attach the volume to the server.

Indeed, it's advised to use the Nova API [2] instead.

This patch doesn't bring any feature, however it brings best practice in
our tests, instead of using "low level" volume APIs, which aren't
advised to our users.

[1] https://docs.openstack.org/python-openstackclient/2023.1/cli/command-objects/volume-attachment.html#volume-attachment-create
[2] https://docs.openstack.org/api-ref/compute/#servers-with-volume-attachments-servers-os-volume-attachments

Fixes #2768