orchestration: Fix relative to abs file refs fixing
Existing fixFileRefs() does a simple string replacement regardless of
context. This is not appropriate as the same keyord could appear outside
of the valid template/file reference cases ('type: ...' and
'get_file: ...').

Example:

  init-cmd:
    type: OS::Heat::CloudConfig
    properties:
      cloud_config:
        write_files:
          - path: /etc/somefile
            content: { get_file: somefile }

Here "somefile" in "{ get_file: somefile }" should be replaced by the
absolute path to somefile, but the "path: /etc/somefile" part should not
be modified as that path refers here to the target destination in a
completely different context (in this example, a cloud-init target path
on a to-be-created VM).

Signed-off-by: David Verbeiren <david.verbeiren@tessares.net>