orchestration: remove fixFileRefs() also for env
An earlier commit removed the usage of fixFileRefs() for templates, by
having the getFileContents() do the replacement directly on the parsed
data structure as it goes through it and gathers referenced files.

This commit does the same for environment files that also ultimately use
the now improved getFileContents() function. As the fixFileRefs() was
not used anywhere else, it has been removed, together with its test.

Signed-off-by: David Verbeiren <david.verbeiren@tessares.net>